+++ common/cpp/src/http_client/curl_http_client.cpp
@@ -222,7 +222,7 @@ std::string CurlHttpClient::UrlEscape(const std::string& uri) const noexcept {
     if (!curl_) {
         return "";
     }
-    char* output = curl_easy_escape(curl_, uri.c_str(), uri.size());
+    char* output = curl_easy_escape(curl_, uri.c_str(), static_cast<int>(uri.size()));
     if (output) {
         auto res = std::string(output);
         curl_free(output);
